Major overhaul of the project Recreated all VS projects, reworked output directories based on platform and configuration Updated gitignore Renamed Map::RemoveKey to Map::RemoveAt Added Utility::Basename Moved Timing from Engine to Common Removed binary files Updated FileLogger, now stores a map of output streams to prevent opening the same file twice. Added Log function code, so FileLogger now works properly.

+++ Common/Tracking/MemoryTracker.h
@@ -30,7 +30,7 @@ public:
 	static bool RemoveAllocation( TrackedObject* pObject );
 
 	static inline unsigned int AllocationCount( void ) 
-		{ return s_Allocations.Size(); }
+		{ return (unsigned int)s_Allocations.Size(); }
 
 	static void PrintAllocations( void );
